Code refactor complete; Functions cleaner/clearer
Separated code based on MVC model. Took a bit of work, but I was able
to simplify a lot of code. I was going on and on about how to send
winner object so that the 'win' class name would be added at the end of
a game. Ended up just sending whoever was playing when the game was won.

Planned features:
    Ability for the system to detect a stalemate and display a message
    Win animations
    A best of three/five option for friends to verse
    Maybe a 1-player (bot) modified
    Fix up the look of everything