Make catscratch more apparent (message box).
So, apparently I had forgotten about adding the system checking for a
tie. It's already in and all I had to do was set up the message box. I
do agree with the sentiment on the last commit:

I THINK THE MESSAGE BOX SHOULD BE REMOVED. It's like calling the user an
idiot like they can't realize who won (especially with the winning
player lighting up at the top as well as their winning squares).
Whatever is done about it, it needs to be fixed. It's HIDEOUS!!!

Planned features:
    A best of three/five option for friends to verse
    Maybe a 1-player (bot) mode