Clean up code. Looks flooky.
The code was cleaned up and made more clear. I don't like the
architecture of the project and I want to redo it at a later point.

I'm on the fence with the message box after coming back to this project
a little later. But I want to retain this message for prosterity:

I THINK THE MESSAGE BOX SHOULD BE REMOVED. It's like calling the user an
idiot like they can't realize who won (especially with the winning
player lighting up at the top as well as their winning squares).
Whatever is done about it, it needs to be fixed. It's HIDEOUS!!!

Planned features:
    A best of three/five option for friends to verse
    Maybe a 1-player (bot) mode